Rename slider input to range input
