+++ projects/packages/forms/src/blocks/contact-form/child-blocks.js
@@ -27,8 +27,8 @@ import JetpackStepContainer from '../form-step-container';
 import JetpackStepDivider from '../form-step-divider';
 import JetpackStepNavigation from '../form-step-navigation';
 import JetpackInput from '../input';
+import JetpackSliderInput from '../input-range';
 import JetpackRatingInput from '../input-rating';
-import JetpackSliderInput from '../input-slider';
 import JetpackLabel from '../label';
 import JetpackOption from '../option';
 import JetpackOptions from '../options';
