+++ projects/packages/forms/src/blocks/input-range/editor.scss
@@ -1,4 +1,4 @@
-.jetpack-input-slider-row {
+.jetpack-input-range-row {
 	display: flex;
 	align-items: center;
 	width: 100%;
@@ -6,20 +6,20 @@
 	margin-top: 30px;
 }
 
-.jetpack-input-slider__min-label,
-.jetpack-input-slider__max-label {
+.jetpack-input-range__min-label,
+.jetpack-input-range__max-label {
 	font-size: .9em;
 	text-align: center;
 }
 
-.jetpack-input-slider__input-container {
+.jetpack-input-range__input-container {
 	display: flex;
 	flex: 1 1 auto;
 	min-width: 0;
 	position: relative;
 }
 
-.jetpack-input-slider__range {
+.jetpack-input-range__range {
 	-webkit-appearance: none;
 	appearance: none;
 	height: 2px;
@@ -31,12 +31,12 @@
 	box-sizing: border-box;
 }
 
-.jetpack-input-slider__range:focus {
+.jetpack-input-range__range:focus {
 	outline: none;
 	opacity: 1;
 }
 
-.jetpack-input-slider__range::-webkit-slider-thumb {
+.jetpack-input-range__range::-webkit-slider-thumb {
 	-webkit-appearance: none;
 	appearance: none;
 	width: 12px;
@@ -48,11 +48,11 @@
 	box-shadow: 0 0 2px rgba(0,0,0,0.2);
 }
 
-.jetpack-input-slider__range:focus::-webkit-slider-thumb {
+.jetpack-input-range__range:focus::-webkit-slider-thumb {
 	background: var(--jetpack--contact-form--button-primary--background-color);
 }
 
-.jetpack-input-slider__range::-moz-range-thumb {
+.jetpack-input-range__range::-moz-range-thumb {
 	width: 12px;
 	height: 12px;
 	background: var(--jetpack--contact-form--button-primary--background-color);
@@ -61,11 +61,11 @@
 	transition: background 0.3s;
 }
 
-.jetpack-input-slider__range:focus::-moz-range-thumb {
+.jetpack-input-range__range:focus::-moz-range-thumb {
 	background: var(--jetpack--contact-form--button-primary--background-color);
 }
 
-.jetpack-input-slider__range::-ms-thumb {
+.jetpack-input-range__range::-ms-thumb {
 	width: 12px;
 	height: 12px;
 	background: var(--jetpack--contact-form--button-primary--background-color);
@@ -74,14 +74,14 @@
 	transition: background 0.3s;
 }
 
-.jetpack-input-slider__range::-ms-fill-lower,
-.jetpack-input-slider__range::-ms-fill-upper,
-.jetpack-input-slider__range:focus::-ms-fill-lower,
-.jetpack-input-slider__range:focus::-ms-fill-upper {
+.jetpack-input-range__range::-ms-fill-lower,
+.jetpack-input-range__range::-ms-fill-upper,
+.jetpack-input-range__range:focus::-ms-fill-lower,
+.jetpack-input-range__range:focus::-ms-fill-upper {
 	background: var(--jetpack--contact-form--button-primary--background-color);
 }
 
-.jetpack-input-slider__value-indicator {
+.jetpack-input-range__value-indicator {
 	position: absolute;
 	top: -36px;
 	left: 0;
