+++ projects/packages/forms/src/blocks/input-range/icon.js
@@ -1,12 +1,12 @@
 import { SVG, Line, Circle } from '@wordpress/components';
 
 /**
- * SliderIcon - A horizontal line with a dot in the middle, for use as a slider field icon.
+ * RangeIcon - A horizontal line with a dot in the middle, for use as a range field icon.
  *
  * @param {object} props - React props
  * @return {Element} The SVG icon
  */
-export default function SliderIcon( props ) {
+export default function RangeIcon( props ) {
 	return (
 		<SVG width="24" height="24" viewBox="0 0 24 24" fill="none" { ...props }>
 			<Line x1="4" y1="12" x2="20" y2="12" stroke="currentColor" strokeWidth="2" />
